Changed WifiCard distance-to-wifi to Int for cleaner UI

+++ app/src/main/java/com/example/wifind/activity/MarketBoardActivity.kt
@@ -126,7 +126,7 @@ class MarketBoardActivity : AppCompatActivity() {
         }
         return WifiCard(
             wifi = this,
-            distanceToWifi = currentLocation.distanceTo(wifiLocation).toDouble()
+            distanceToWifi = currentLocation.distanceTo(wifiLocation).toInt()
         )
     }
