+++ app/src/main/java/com/example/wifind/model/WifiCard.kt
@@ -2,5 +2,5 @@ package com.example.wifind.model
 
 data class WifiCard(
     val wifi: Wifi,
-    val distanceToWifi: Double
+    val distanceToWifi: Int
 )
